+++ data_display/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class DatadisplayConfig(AppConfig):
+    name = 'data_display'
