+++ data_display/urls.py
@@ -1,0 +1,9 @@
+from django.conf.urls import url
+
+from . import views
+from django.urls import path
+
+urlpatterns = [
+    path('', views.database_start_page, name='database_start_page'),
+    path('', views.display_data, name = "display_data" )
+]
