--- datadisplay/static/datadisplay/style_site.css
@@ -1,55 +1,0 @@
-table {
-  font-family: arial, sans-serif;
-  border-collapse: collapse;
-  width: 100%;
-}
-
-td, th {
-  border: 3px solid rgb(175, 76, 170);
-  text-align: left;
-  padding: 10px;
-}
-
-tr:nth-child(even) {
-  background-color: rgba(221, 221, 221, 0.692);
-}
-
-h1 {
-  right: 50px;
-  left: 50 px;
-  padding: 12px 20px;
-  background-color: rgb(175, 76, 170);
-  display: inline-block;
-}
-
-h2, p {
-  right: 50px;
-  left: 50px;
-  padding: 12px 20px;
-  display: inline-block
-}
-
-input[type=text], select {
-    width: 100%;
-    padding: 12px 20px;
-    margin: 9px 0;
-    display: inline-block;
-    border: 1px solid rgb(175, 76, 170)c;
-    border-radius: 4px;
-    box-sizing: border-box;
-  }
-  
-input[type=submit] {
-   width: 100%;
-   background-color: rgb(175, 76, 170);
-   color: white;
-   padding: 14px 20px;
-   margin: 8px 0;
-   border: none;
-   border-radius: 4px;
-   cursor: pointer;
- }
-
-input[type=submit]:hover {
-  background-color: rgb(200, 76, 170);
- }
