+++ veep_data_project/settings.py
@@ -31,7 +31,7 @@ ALLOWED_HOSTS = []
 # Application definition
 
 INSTALLED_APPS = [
-    'datadisplay.apps.DatadisplayConfig',
+    'data_display.apps.DatadisplayConfig',
     'django.contrib.admin',
     'django.contrib.auth',
     'django.contrib.contenttypes',
