+++ veep_data_project/urls.py
@@ -18,5 +18,5 @@ from django.urls import path, include
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('datadisplay/', include('datadisplay.urls')),
+    path('data_display/', include('data_display.urls')),
 ]
